Get Documented Functions in a Package

Lists all documented functions in a package together with some of their documentation elements as raw text. Only works for installed packages.

Usage

get_pkg_docs(pkg = NULL, unload = TRUE, reload = TRUE)

Arguments

pkg

The package name. If NULL, the package name is inferred from the DESCRIPTION file in the current directory or any parent directory. If no DESCRIPTION file is found, the function stops with an error message.

unload

Whether to unload a potential currently developed package using devtools::unload() before checking the documentation. Required when the package was loaded with devtools::load_all() as the documentation database only exists for installed packages.

reload

Whether to reload the package using devtools::load_all() after checking the documentation.

Value

Returns a dataframe with columns title, description, value, examples and rows corresponding to the documented functions in the package.

Examples

df <- get_pkg_docs("tools")
nchars <- as.data.frame(apply(df, 2, function(col) sapply(col, nchar)))
print(nchars)
#>                                              class title description value
#> HTMLheader                                       8    42          71   114
#> Rcmd                                             8    15          33    30
#> Rd2HTML                                          8    15         181   644
#> Rd2txt_options                                   8    36          61   154
#> RdTextFilter                                     8    25          87   128
#> Rdiff                                            8    25          99   232
#> Rdindex                                          8    28         221     0
#> SweaveTeXFilter                                  8    31         120   138
#> add_datalist                                     8    40         284     0
#> assertCondition                                  8    26         423   284
#> bibstyle                                         8    37         131   141
#> buildVignette                                    8    18         162    61
#> buildVignettes                                   8    32         111   240
#> checkFF                                          8    28         299   120
#> checkMD5sums                                     8    35          50    74
#> checkPoFiles                                     8    55         209   144
#> checkRd                                          8    20          59   928
#> checkRdaFiles                                    8    51         250   556
#> checkTnF                                         8    32         354   317
#> checkVignettes                                   8    23         267   215
#> check_packages_in_dir                            8    52          88     0
#> codoc                                            8    36         311   571
#> compactPDF                                       8    18         111   255
#> delimMatch                                       8    26          71   237
#> dependsOnPkgs                                    8    25         124    75
#> find_gs_cmd                                      8    43          68   126
#> getVignetteInfo                                  8    38          55   290
#> loadRdMacros                                     8    39          81   270
#> makeLazyLoading                                  8    24          52     0
#> make_translations_pkg                            8    48          75     0
#> matchConcordance                                 8    44         500   341
#> md5sum                                           8    21          53   213
#> package_dependencies                             8    52          70   311
#> package_native_routine_registration_skeleton     8    67          70    51
#> parseLatex                                       8    46         259   384
#> parse_Rd                                         8    16          99   518
#> pkg2HTML                                         8    15         112    41
#> pskill                                           8    14          61    84
#> psnice                                           8    47          80    77
#> showNonASCII                                     8    29         126    78
#> startDynamicHelp                                 8    34         100    95
#> testInstalledPackage                             8    23          95   245
#> texi2dvi                                         8    19         112   159
#> toHTML                                           8    25          69   119
#> toRd                                             8    59         138   288
#> toTitleCase                                      8    28          69    62
#> undoc                                            8    25         194   242
#> update_pkg_po                                    8    35          91     0
#> vignetteEngine                                   8    39         563   230
#> vignetteInfo                                     8    34          46   391
#> xgettext                                         8    55         420   262
#>                                              format examples
#> HTMLheader                                        0       53
#> Rcmd                                              0        0
#> Rd2HTML                                           0      321
#> Rd2txt_options                                    0      205
#> RdTextFilter                                      0        0
#> Rdiff                                             0        0
#> Rdindex                                           0        0
#> SweaveTeXFilter                                   0        0
#> add_datalist                                      0        0
#> assertCondition                                   0     1303
#> bibstyle                                          0      854
#> buildVignette                                     0        0
#> buildVignettes                                    0       70
#> checkFF                                           0       74
#> checkMD5sums                                      0        0
#> checkPoFiles                                      0       49
#> checkRd                                           0      361
#> checkRdaFiles                                     0      264
#> checkTnF                                          0        0
#> checkVignettes                                    0        0
#> check_packages_in_dir                             0      898
#> codoc                                             0        0
#> compactPDF                                        0        0
#> delimMatch                                        0       81
#> dependsOnPkgs                                     0      187
#> find_gs_cmd                                       0      150
#> getVignetteInfo                                   0       26
#> loadRdMacros                                      0      343
#> makeLazyLoading                                   0        0
#> make_translations_pkg                             0        0
#> matchConcordance                                  0        0
#> md5sum                                            0       71
#> package_dependencies                              0     1008
#> package_native_routine_registration_skeleton      0      726
#> parseLatex                                        0       70
#> parse_Rd                                          0        0
#> pkg2HTML                                          0       70
#> pskill                                            0       29
#> psnice                                            0        0
#> showNonASCII                                      0      224
#> startDynamicHelp                                  0        0
#> testInstalledPackage                              0      328
#> texi2dvi                                          0        0
#> toHTML                                            0       48
#> toRd                                              0        0
#> toTitleCase                                       0        0
#> undoc                                             0       66
#> update_pkg_po                                     0        0
#> vignetteEngine                                    0       30
#> vignetteInfo                                      0       94
#> xgettext                                          0     1340